# Add a spin before the cond_wait in the thread pool