Opening for discussion. This improves performance on a few apps by 1-3%, and improves the performance of performance_inner_loop_parallel by more like 20%.

@zvookin