@abadams 2020-10-26
This was motivated by noticing that our cond_wait in synchronization_common.h doesn't spin before sleeping.