@zvookin 2020-10-27
Does the total number of CPU cycles go up? I would expect it to, but the question is whether it goes up a noticeable amount or just a little bit. Have you looked at how it interacts with the atomic index update change especially for the inner loop case. I'll check it on the benchmarking cases I've been looking at.

The only significant question is whether the experiment is good enough to prove it is a win enough of the time and not a big lose in any case we care about. (Which seems unlikely given 40 spins just isn't that many and it should not be happening particularly often. Like unless there is a bug or my quick analysis is way off.) Then again, the 40 is sort of becoming this magical number that isn't really well grounded so...

Not objection to landing it pending a bit more experimentation, but it strikes me as more a sign that our design is fragile and could be improved than a principled approach. Longer term, I think having the owners sleep on individual condvars and arranging for targeted thread handoffs is probably the better way to get deterministic and closer to optimal scheduling.