@abadams 2020-10-27
CPU cycles goes up by about 1%, runtime of good schedules goes down by about 1%. Oddly though I'm seeing the runtime of some schedules that use nested parallelism go up by about 1%, and I haven't been able to figure out why. I've verified that those schedules don't spin anyway (because there are always more jobs on the job stack).