@abadams 2020-11-06
Updated to also spin on mutexes even if threads are parked, and performance of the autoschedule (which uses fine-grained parallelism) went from 53ms to 35ms! Similar benefit in interpolate app: 32ms -> 18ms