@zvookin 2020-11-10
We need to move all the par for experiments into a single branch and do some experiments to evaluate them. My running of the first version of this change is that it barely helps and makes a couple things slightly worse. I need to pickup the latest, but I'd like all of this to end up in the par_for_opt branch and to have spreadsheets showing the measurements of the stuff we're looking at across a chosen set of conditions. Otherwise we're playing whackamole.