@abadams 2021-02-01
I think we should just land it. It's a minor change with a pretty good performance boost on some apps (note that Zalman's comment applied to the first version of the change).