@zvookin 2021-02-01
tl;dr: Feel free to land it but I'm pretty sure it is a lucky thing that won't stand the test of time.

My concern is that in the grand scheme of things, this is more noise than signal. It is another chunk of complexity in a subsystem that needs to be approached more methodically and after doing so, will likely not have this code in place. One can argue that a short term performance improvement is worth it and I don't much disagree. But to make that case reliably, we need performance and resource usage across a broader class of uses. (E.g. it likely depends a lot on the number of cores in the hardware, etc.) It would be really nice if the performance data we do have was in a table in the PR description to help with future work.

Approaching things more methodically involves creating a set of benchmarks that measure at least time and resource usage across a variety of work item and loop length sizes, then picking at least a few points in the hardware space. We should probably also try to have a mechanism to get feedback from users, such as some way to use Google's large scale profiling infrastructure to track the benefit (or cost) of changes like this over time.

My guess is the coding approach will evolve both the threadpool and fast sync mechanisms somewhat. E.g. moving to a conditional critical section design, instead of mutes/cond_wait, which allows pushing the condition evaluation into the fast sync scheduling code and eliminating some amount of spurious wakeups. We've also talked about a signal N threads mechanism to avoid the A/B team logic. (Conditional critical section probably takes care of that too.) The fundamental algorithms of how we do e.g. a parallel for loop also need investigation and likely need to become adaptive.