@abadams 2021-02-01
Yeah, I don't think it solves all our problems, but it's sane, and it improve things in some situations. I'll rerun all the apps at least and report back.