@abadams 2021-02-01
app | speedup | before | after
-- | -- | -- | --
bilateral manual | 1.002274472 | 5.00152 | 4.99017
bilateral auto | 1.005950422 | 5.16633 | 5.13577
bgu manual | 1.006090277 | 8.85121 | 8.79763
bgu auto | 1.189101281 | 31.4767 | 26.471
conv manual | 0.9863711133 | 4.79113 | 4.85733
conv auto | 0.9998220172 | 12.9203 | 12.9226
hist manual | 1.02685121 | 1.12547 | 1.09604
hist auto | 1.413868213 | 7.45175 | 5.27047
interpolate manual | 1.014333005 | 4.51648 | 4.45266
interpolate auto | 1.24461596 | 16.3031 | 13.0989
lens blur manual | 1.026810895 | 4.42728 | 4.31168
lens blur auto | 1.114805521 | 12.439 | 11.158
local laplacian manual | 0.9972458769 | 24.8033 | 24.8718
local laplacian auto | 1.063665744 | 34.8208 | 32.7366
max filter manual | 1.012758825 | 48.6423 | 48.0295
max filter auto | 1.021298673 | 67.5202 | 66.1121
nl means manual | 1.000094986 | 69.4907 | 69.4841
nl means auto | 1.000902275 | 65.893 | 65.8336
stencil chain manual | 1.038385325 | 4.44377 | 4.2795
stencil chain auto | 0.993515189 | 9.42986 | 9.49141
geomean: | 1.053347848 |   |  