@abadams 2021-02-01
Something not captured in the stats is that this seems to make runtimes a lot less variable.