@abadams 2021-02-01
On the auto schedule for hist, the standard deviation of runtimes goes from 0.85ms down to 0.29ms over 1000 runs (relative to an average runtime of ~6ms)