@zvookin 2021-02-01
Please add a brief description of the hardware. (Instruction set, num cores, rough clock speed, whether variable clocks are enabled.)